Add `get_bosses` to `Subscription` (#10)
Subscriptions now have a `get_bosses` method that will request a
`BossList` message to be sent to the subscriber. Since this boss list
doesn't change often, we cache it to avoid having to call `map_message`
multiple times needlessly if the list hasn't changed.

Every time the boss list is changed (boss is added or changed), the
worker calls the `map_message` function on the boss list and caches the
result. The subscriber gets this cached result when calling
`get_bosses`.